Refactor the resolveParser to return an array of parsers.
This allows us to not create multiple parsers for the `nextText`, we just resolve one parser for the type and optionally resolve for nextText. This makes things faster by another order of magnitude.

This design was also demanded due to the casing rules for `Lowercase`, `Uppercase`, `Uncapitalize`, and `Capitalize`.

Also fixes the stack overflow when transforming a type into its base constraint when there's actually no difference between the constraints and the types in TemplateLiteralType.