Implement a between parser.
This helps to fix the case where, given the type `` `${number}.${number}.${number}` ``, the input `1.1.1` would consume `1.1` for the first `number`, `.` is then consumed, leaving the final `1` to be consumed by the second `number`. This fails because the cursor didn't reach the end.

The between parser will advance in the same way that `skipUntil` does, but moves the cursor back, and invokes the parser before the original parser and makes sure that's sound.

As implemented, it's not quite perfect either. Maybe we need to create a slice of the cursor and then invoke the parser on that instead?