Undo the optimization for the or parser, sadly.
If I have a union `"http" | "https"` and I expect a `://`, then given the literal `"https://"`, this fails because `"http"` is a substring of `"https"` and is taken, which places the cursor right where the character `s` is. Then I expect a `://` to follow, and that fails, because `s` is not `:`.

Now the parser has to work slightly different, e.g. we now expect, in a pair, the type in the union and the next parser both must succeed.

That means we require `"http"` to succeed if `"://"` also succeeds right after it. If it does not, then backtrack and try `"https"` followed by `"://"`.